requirements.txt
pymongo~=4.6.3
python-dotenv~=1.0.1
holidays~=0.47
numpy~=1.26.4
pandas~=2.2.2
seaborn~=0.13.2
matplotlib~=3.8.4
scikit-learn~=1.4.2
motor~=3.4.0
pyspark~=3.5.1
requests
pyyaml
xgboost